job-list: move job struct into its own files
Problem: In the future, additional services added into job-list
may lead to header circular includes.

Solution: Move struct job and job create/destroy functions into
their own files to break future circular dependencies.
