+++ src/modules/job-list/job_state.c
@@ -28,6 +28,7 @@
 #include "src/common/libidset/idset.h"
 
 #include "job_state.h"
+#include "job_data.h"
 #include "idsync.h"
 #include "job_util.h"
 
@@ -114,59 +115,10 @@ static int job_inactive_cmp (const void *a1, const void *a2)
     return NUMCMP (j2->t_inactive, j1->t_inactive);
 }
 
-static void job_destroy (void *data)
-{
-    struct job *job = data;
-    if (job) {
-        free (job->ranks);
-        free (job->nodelist);
-        json_decref (job->annotations);
-        grudgeset_destroy (job->dependencies);
-        json_decref (job->jobspec);
-        json_decref (job->R);
-        free (job->eventlog);
-        json_decref (job->exception_context);
-        zlist_destroy (&job->next_states);
-        free (job);
-    }
-}
-
 static void job_destroy_wrapper (void **data)
 {
     struct job **job = (struct job **)data;
     job_destroy (*job);
-}
-
-static struct job *job_create (struct list_ctx *ctx, flux_jobid_t id)
-{
-    struct job *job = NULL;
-
-    if (!(job = calloc (1, sizeof (*job))))
-        return NULL;
-    job->ctx = ctx;
-    job->id = id;
-    job->userid = FLUX_USERID_UNKNOWN;
-    job->urgency = -1;
-    /* pending jobs that are not yet assigned a priority shall be
-     * listed after those who do, so we set the job priority to MIN */
-    job->priority = FLUX_JOB_PRIORITY_MIN;
-    job->state = FLUX_JOB_STATE_NEW;
-    job->ntasks = -1;
-    job->nnodes = -1;
-    job->expiration = -1.0;
-    job->wait_status = -1;
-    job->result = FLUX_JOB_RESULT_FAILED;
-
-    if (!(job->next_states = zlist_new ())) {
-        errno = ENOMEM;
-        job_destroy (job);
-        return NULL;
-    }
-
-    job->states_mask = FLUX_JOB_STATE_NEW;
-    job->states_events_mask = FLUX_JOB_STATE_NEW;
-    job->eventlog_seq = -1;
-    return job;
 }
 
 /* zlistx_insert() and zlistx_reorder() take a 'low_value' parameter
